# This PR significantly speeds up lossless decoding. 