Dear authors, your code really helped us out, and we want to contribute some of our improvements.

In my tests on a bit CT volume the speedup is 12 sec -> 4.5 sec. Roughly 2.5x.

The unit tests are all passing. Unit test time is reduced 850 ms -> 66 ms.

Speedup is obtained by replacing DataView by native Int8Array / Uint8Array / Int16Array.