@yvoronen 2016-10-28
Note, that I tested on latest Chrome browser.